requirements.txt
Flask==2.3.3
Werkzeug==2.3.6
python-dotenv==1.0.1
cloudinary==1.38.0
gunicorn==21.2.0
